Rename components with capital letter, create RecipeCard
