+++ frontend/src/components/StarRating.jsx
